Added more metrics. Mem values currently wrong

+++ raw-metrics.go
@@ -10,44 +10,44 @@ import (
 // InstanceMetrics models 1 of RawMetrics Instances collection
 type InstanceMetrics struct {
 	InstanceSuffix             string
-	RestartCounter             int
-	UserRestartCounter         int
-	Uptime                     uint64
+	RestartCounter             float64
+	UserRestartCounter         float64
+	Uptime                     float64
 	InUse                      bool
-	UseCounter                 int
+	UseCounter                 float64
 	ModuleStatus               string
 	ProcessName                string
-	CommittedVirtualMemorySize uint64
-	ProcessCPUTime             uint64
-	HeapCommittedMemorySize    uint64
-	HeapUsedMemorySize         uint64
-	NonHeapCommittedMemorySize uint64
-	NonHeapUsedMemorySize      uint64
+	CommittedVirtualMemorySize float64
+	ProcessCPUTime             float64
+	HeapCommittedMemorySize    float64
+	HeapUsedMemorySize         float64
+	NonHeapCommittedMemorySize float64
+	NonHeapUsedMemorySize      float64
 }
 
 // RawMetrics models the result of the metrics request to the PTV server
 type RawMetrics struct {
 	ServiceName                string
 	MinPoolSize                float64 // TODO: Make all numbers compatible with Prometheus types
-	MaxPoolSize                int
-	NumSuccess                 uint64
-	NumFailure                 uint64
-	NumRejected                uint64
-	NumService                 int
-	AvgInnerTime               int
-	MinInnerTime               int
-	MaxInnerTime               int
-	AvgComputationTime         int
-	AvgOuterTime               int
-	MinOuterTime               int
-	MaxOuterTime               int
+	MaxPoolSize                float64
+	NumSuccess                 float64
+	NumFailure                 float64
+	NumRejected                float64
+	NumService                 float64
+	AvgInnerTime               float64
+	MinInnerTime               float64
+	MaxInnerTime               float64
+	AvgComputationTime         float64
+	AvgOuterTime               float64
+	MinOuterTime               float64
+	MaxOuterTime               float64
 	ProcessName                string
-	CommittedVirtualMemorySize uint64
-	ProcessCPUTime             uint64
-	HeapCommittedMemorySize    uint64
-	HeapUsedMemorySize         uint64
-	NonHeapCommittedMemorySize uint64
-	NonHeapUsedMemorySize      uint64
+	CommittedVirtualMemorySize float64
+	ProcessCPUTime             float64
+	HeapCommittedMemorySize    float64
+	HeapUsedMemorySize         float64
+	NonHeapCommittedMemorySize float64
+	NonHeapUsedMemorySize      float64
 	Instances                  []InstanceMetrics
 }
